+++ src/app/Palette.js
